it turns out I REALLY want to kill whoever designed cube maps, I hate them, they never fucking work and are a pain in the ass

+++ src/BluminEngine5/Rendering/Texture.java
@@ -28,7 +28,7 @@ import static org.lwjgl.opengl.GL14.GL_MIRRORED_REPEAT;
 public class Texture implements Serializable{
 
     private org.newdawn.slick.opengl.Texture texture;
-    private int width, height;
+    private float width, height;
     private ByteBuffer decodedbytes;
     private int textureId;
     public ArchivedFile file;
@@ -51,8 +51,8 @@ public class Texture implements Serializable{
         if(file != null) {
             try{
                 texture = GetTexFromFile();
-                width = (int)texture.getWidth();
-                height = (int)texture.getHeight();
+                width = texture.getWidth();
+                height = texture.getHeight();
                 textureId = texture.getTextureID();
 
                 decodedbytes = ByteBuffer.allocate(file.GetDecodedData().length * 4);
@@ -68,8 +68,8 @@ public class Texture implements Serializable{
                 InputStream is = Utils.LoadFileAsStream(fileBackup);
                 File f = new File(fileBackup);
                 texture =  TextureLoader.getTexture(FilenameUtils.getExtension(fileBackup), is , GL11.GL_LINEAR);
-                width = (int)texture.getWidth();
-                height = (int)texture.getHeight();
+                width = texture.getWidth();
+                height = texture.getHeight();
                 textureId = texture.getTextureID();
                 byte[] data = texture.getTextureData();
 
@@ -91,11 +91,11 @@ public class Texture implements Serializable{
         GL13.glDeleteTextures(textureId);
     }
 
-    public int getWidth() {
+    public float getWidth() {
         return width;
     }
 
-    public int getHeight() {
+    public float getHeight() {
         return height;
     }
